chore: Update OpenVPN Docker deployment script to ensure easy-rsa.vars exists
The OpenVPN Docker deployment script has been updated to ensure that the easy-rsa.vars file exists in the /etc/openvpn/config directory. This file contains the configuration variables for the Easy-RSA tool used by OpenVPN. By including this file, it simplifies the setup process and ensures that the necessary variables are set correctly.